feat(util): detect webp support
